Initialize consumers in output datastream as part of parent consumer
This should avoid races between initializers and row processing.

This also extends testing of the reader to ensure that a built job is also runnable. This tests more than the reader, so it might not be the optimal place, but I'm not really sure where it would be, as it is more of an "internal" integration test, testing that the parts assembling and running a job work properly together.

I'll also be lazy and put a small fix of CompletenessAnalyzer on here: There was absolutely no need to create a dataset header, and it actually introduced a bug.