Avoid potentially losing lines
This is more of a workaround than a proper fix, but a proper fix is a bit limited by our interface. The main issue is that the property widgets all have their own listeners directly connected to the builders, making it hard to control listeners without yanking the entire dialog window (which is actually still not bullet-proof). A proper fix would probably only allow (or strongly encourage) listeners to listen to their dialog, which would then have a general component listener (also not currently supported) on the component and/or analysisJobListener.

Luckily, only MultipleInputColumnsPropertyWidget seems to have been directly looking at the input from the listener, instead of just updatingfrom the general state, which should make the other TransformerChangeListener-implementing PropertyWidgets immune. Heavy testing is still recommended, though.

Fixes #667
Also contains a fix for re-adding listeners, as well as a wrong addTransformerChangeListener on panel close (none of them actually .