Try to update scope for filters
Filter requirements are now treated like other components, so a component will have its scope changed if a filter from another scope is linked to it.

This also adds a general confirmation for any scope changes that will destroy user configuration, as well as remove a bit of duplication.

Fixes #909