Issue #967: Fixed
The issue was caused by the NumberFormat applying grouping (thousand
separator) which then made the FormattedTextField try to submit the
value as e.g. "8,020" instead of "8020". This then caused the
NumberDocument to reject the value and return empty string instead.

Also did a small change to update the default port from 9000 to 8020.