Partial fix for preview filter issue
Issue was that max rows filter was added using OR compound filter, meaning that all other filters was essentially nuked (as well as the max rows filter if other filters were).

I extended the filter to also support ALL scenario, though it is a bit of weird hybrid now (normally ALL scenario is done by chaining requirements).

For some reason, this does NOT fix the test case, allthough it fixed a manual test in DC (I used the first example Kasper provided in fixed case).

This fix is obviously not mergeable with a broken unit test, but I'd like some feedback to my approach.

Fixes #1047