Yet another attempt to migrate away from Scala results
This latest attempt tries to leave almost everything but the actual results in place. Tests are almost untouched. Comments are more than welcome (needed, actually)

A common interface is made for all result classes, the old Scala classes are kept in place, but with their interfaces changed to match the interface. A new Java-based implementation is then added, and analyzers changed to use those. There are lots of conversions going on, but most of these should be relatively cheap, as a thin wrapper is used. The exception is the loaded compatibility classes, but those are only used for reading and the conversion is cached, so the cost should be acceptable.

One worry which needs to be verified: To avoid even more conversions, the ScatterAnalyzerResult now uses the Java interface in its initializer, which could be an issue, depending on how that is really implemented.

Please note that the conversion is not done, and barely tested, except for the original tests. Still needs the StackedAreaAnalyzerResult, as well as extra tests for deserialization of both types.

Fixes #1268
Hopefully obsoletes #1294 and #1297 (But lets not remove the underlying branches just yey. Both have started moving the actual analyzers and renderers to Java, which we want to do soon-ish)