Adds and fixes deserialization tests for all three visualizations
This one was a bit of a bastard: We don't really support renaming classes with inner classes very well, and Scala makes a ton of them. In the end I had to give up on changing the name of the Scala ScatterGroup, and instead use a nasty like "CommonScatterGroup" for the interface.

For some reason, I also had to override the autogenerated serial id for that class, even though it wasn't necessary for the result classes. That worries me, so we may want to do pretty thorough testing