Refactored job triggering web service to consume multipart/form-data so you no longer post a path to a properties file, but you actually post a file (a properties file that is). As a result of this, instead of passing the path of the file through in different method calls, now we pass a Map of the properties in the properties file. I choose to convert the Properties object to a Map in the JobTriggeringController to avoid problems with google.gwt.
Also removed the overloaded getAnalyzerBeansConfiguration method from ConfigurationCache, as it doesn't cache the DataCleanerConfiguration it produces (and it had no reason to based its arguments). Added a package protected readConfiguration method instead, which is only called from the TenantContextImpl. Because the ConfigurationCache has an InjectionManagerFactory as a field member it seemed more logic to create an overloaded method on the ConfigurationCache, then to create something similar on the TenantContextImpl.