Use proper parsing instead of non-working homegrown solution
This many issues with the standardizer, but also breaks compatibility a little bit, as it does not add username to host, or fragment to query. I did keep the a field with no path at all as null, as it is not technically filled. I had a bit of a hard time deciding here, so please comment if you disagree.

 For the sake of not breaking jobs, the port number is kept a string even though that makes no sense.

 Fixes #1490