Stop inferring custom column names (#1618)
* Stop inferring custom column names

In case no column names were given, we tried to infer what they would be. Unfortunately that didn't work as expected, instead throw an exception, so we're removing again. This might need rethinking when we're actually going to use that data, but this was only a preparation for the new MM.

Fixes #1617
Fixed NCIM/identify7#317

* Unused import

(cherry picked from commit 839e174)